requirements.txt
flask~=2.0.0
web3==5.22.0
simplejson~=3.17.2
python-dotenv~=0.21.0
marshmallow~=3.19.0